Move OSUtils to root package
